LayoutTests/platform/ios-sim/iphone/scrolling/overflow-div-scrolling-expected.txt

Tests that touch events cause an overflow:scroll div to scroll.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS targetDiv.scrollLeft is 0
PASS targetDiv.scrollTop is 0
PASS targetDiv.scrollLeft is 0
PASS targetDiv.scrollTop is 100
PASS targetDiv.scrollLeft is 0
PASS targetDiv.scrollTop is 50
PASS targetDiv.scrollLeft is 100
PASS targetDiv.scrollTop is 150
